Valuelife/requirements.txt
absl-py==0.9.0
aiodns==2.0.0
aiohttp==3.7.3
aiomultiprocess==0.8.0
aiosqlite==0.16.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.3.0
backoff==1.10.0
beautifulsoup4==4.9.3
bleach==1.5.0
Brlapi==0.6.6
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.1
censys==1.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
colorama==0.4.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.2.10
dnspython==2.1.0
docutils==0.17.1
firebase==3.0.1
firebase-admin==4.5.3
google-api-core==1.26.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.1.0
h5py==2.10.0
html5lib==0.9999999
httplib2==0.19.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.1.1
importlib-resources==5.1.0
ipython==5.5.0
ipython-genutils==0.2.0
joblib==0.15.1
Keras==2.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
Kivy==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.6.2
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.2.1
MouseInfo==0.1.3
msgpack==1.0.2
multidict==5.1.0
netaddr==0.8.0
netifaces==0.10.4
numpy==1.18.5
oauth==1.0.1
oauthlib==3.1.1
olefile==0.45.1
packaging==20.9
pandas==1.0.4
pexpect==4.2.1
phonenumbers==8.12.15
pickleshare==0.7.4
Pillow==8.1.2
plotly==4.14.3
prompt-toolkit==1.0.15
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycairo==1.16.2
pycares==3.1.1
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyee==8.1.0
PyGetWindow==0.0.9
Pygments==2.2.0
pygobject==3.26.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.1.2
pyparsing==2.4.7
pyperclip==1.8.1
pyppeteer==0.2.5
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
PyRect==0.1.4
pyRFC3339==1.0
PyScreeze==0.1.26
PySocks==1.7.1
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python-dotenv==0.19.1
python3-xlib==0.15
PyTweening==1.0.3
pytz==2021.1
pywhatkit==3.8
pyxdg==0.25
PyYAML==5.4.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
reportlab==3.4.0
requests==2.25.1
requests-futures==1.0.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.23.1
scipy==1.4.1
SecretStorage==2.3.1
selenium==3.141.0
shodan==1.25.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.15.0
sklearn==0.0
soupsieve==2.2
sqlparse==0.4.2
stem==1.8.0
system-service==0.3
systemd-python==234
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
texttable==1.6.3
threadpoolctl==2.1.0
timg==1.1.5
torrequest==0.1.0
tqdm==4.56.0
traitlets==4.3.2
typing==3.7.4.3
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.4
usb-creator==0.3.3
uvloop==0.14.0
wadllib==1.3.2
wcwidth==0.1.7
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.3.0
wikipedia==1.4.0
xkit==0.0.0
XlsxWriter==1.3.7
yarl==1.6.3
zipp==3.4.0
zope.interface==4.3.2
